Keep context definition and provider implementation in a single file
