+++ app/src/editor/canvas/components/CanvasController/useCanvasLoadCallback.js
@@ -1,6 +1,6 @@
 import { useContext, useCallback } from 'react'
 
-import RouterContext from '$shared/contexts/Router'
+import { Context as RouterContext } from '$shared/components/RouterContextProvider'
 import useIsMountedRef from '$shared/utils/useIsMountedRef'
 import usePending from '$editor/shared/hooks/usePending'
 import { Context as PermissionContext } from '$editor/canvas/hooks/useCanvasPermissions'
